Adding MATLAB code for Huffman encoding.
huffman_code and split_tree are used to assemble the Huffman dictionary
and encode the image. It currently does not write it to a binary file,
but returns it as an ASCII string of zeros and ones, as MATLAB does not
support binary numbers natively.

Unwrap_huffman runs, albeit slowly.